Update source to compile with HIP + fix some bugs.
Bug #1:

Fix streams in state_generator

We found that the streams are not necessarily created on the device
on which they will be used. Therefore, we set the correct device
with device->set_active before creating the state (which in turn creates the
stream). This also seems to fix timeout warnings we have seen in some benchmarks.

Bug #2:

Delete move assignment as default assignment cannot be generated by hipcc

Bug #3:

Fixed markdown output for HIP + fixed a few compiler warnings