Drop support for CTK < 11.8 and clang < 14.
Newer versions of fmt have a ton of issues building on CTK 11.1, and 11.8 is the next available container we have built for CI. We may still work with some earlier versions, but we do not test them.

We no longer have CI images available for clang < 14, so drop official support.